+++ lib/src/analyzers/lint_analyzer/rules/node_utils.dart
@@ -1,0 +1,15 @@
+import 'package:analyzer/dart/ast/ast.dart';
+
+String humanReadableNodeType(AstNode? node) {
+  if (node is ClassDeclaration) {
+    return 'Class';
+  } else if (node is EnumDeclaration) {
+    return 'Enum';
+  } else if (node is ExtensionDeclaration) {
+    return 'Extension';
+  } else if (node is MixinDeclaration) {
+    return 'Mixin';
+  }
+
+  return 'Node';
+}