+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_match_file_name/examples/multiple_enums.dart
@@ -1,0 +1,5 @@
+enum _MultiEnum { a, b, c }
+
+enum MultipleEnumExample { a, b, c }
+
+enum Test { a, b, c }