+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_match_file_name/examples/multiple_extensions.dart
@@ -1,0 +1,5 @@
+extension _MultiExtension on String {}
+
+extension MultipleExtensionExample on String {}
+
+extension Test on String {}